test i2c update

+++ I2C_protocol/I2CBridge.cpp
@@ -127,7 +127,7 @@ int main(int argc, char *argv [])
 				SwitchServer* s = (SwitchServer*) ImplementedResourceTypes::createServerOfType(OIC_SWITCH_TYPE, name);
 				s->setI2CDevice(uno3,res.first);
 				cout << uno3.readData(res.first) << endl;
-				cout << uno3.getResourceLogic(res.first) << endl;
+				cout << (int) uno3.getResourceLogic(res.first) << endl;
 				s->setInitialState(uno3.readData(res.first) == 1);
 				servers.push_back(s);
 				break;
